+++ src/edge-agent/helpers/ProtocolHelpers.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { AttachmentDescriptor, Message } from "../../domain";
 import { AgentError } from "../../domain/models/Errors";
-import { isArray, isEmpty, isNil, notEmptyString } from "../../utils";
+import { isNil } from "../../utils";
 import { CredentialFormat } from "../protocols/issueCredential/CredentialFormat";
-import { MediationGrantBody, BasicMessageBody, ProblemReportBody } from "../protocols/types";
+import { MediationGrantBody } from "../protocols/types";
 
 export const parseCredentialAttachments = (credentials: Map<string, any>) => {
   const initialValue = {
@@ -27,30 +27,6 @@ export const parseCredentialAttachments = (credentials: Map<string, any>) => {
     initialValue
   );
 };
-
-export const parseBasicMessageBody = (msg: Message): BasicMessageBody => {
-  if (notEmptyString(msg.body.content)) {
-    return {
-      content: msg.body.content,
-    };
-  }
-
-  throw new AgentError.InvalidBasicMessageBodyError("Invalid content");
-};
-
-export const parseProblemReportBody = (msg: Message): ProblemReportBody => {
-  if (notEmptyString(msg.body.code) &&
-    notEmptyString(msg.body.comment) &&
-    notEmptyString(msg.body.escalate_to) &&
-    isArray(msg.body.args) && isEmpty(msg.body.args)
-  ) {
-    const { code, comment, escalate_to, args } = msg.body;
-    return { code, comment, escalate_to, args };
-  }
-  throw new AgentError.InvalidProblemReportBodyError();
-};
-
-
 
 export const parseMediationGrantMessage = (msg: Message): MediationGrantBody => {
   if (isNil(msg.body.routing_did)) {
