+++ src/edge-agent/protocols/types.ts
@@ -28,27 +28,8 @@ export interface PickupAttachment {
   data: string;
 }
 
-export interface PrismOnboardingInvitationBody {
-  type: string;
-  onboardingEndpoint: string;
-  from: string;
-}
-
-export interface BasicMessageBody {
-  content: string;
-}
-
-export interface ProblemReportBody {
-  code: string,
-  comment: string,
-  args: string[],
-  escalate_to: string;
-}
-
 export interface HandshakeRequestBody {
   goalCode?: string;
   goal?: string;
   accept?: string[];
 }
-
-export type OutOfBandInvitationBody = HandshakeRequestBody;
